+++ NEWS.md
@@ -4,6 +4,8 @@ pavo 1.3.2 (development)
 MINOR FEATURES AND BUG FIXES 
 
 * fixed bug in voloverlap() when plot = TRUE
+* fixed bug in vismodel() when transmission has more than one column
+* fixed bug in vismodel() that applied von Kries correction to achromatic channel
 * added argument fill=FALSE in voloverlap()
 
 pavo 1.3.1